fix: avoid float-to-double promotion warning in block norm visitors
nanobind::float_ only has a double constructor, so passing a plain
float widened it implicitly and tripped -Wdouble-promotion under
strict Clang builds. Cast to double explicitly instead.

+++ python/blocknormalizer.hpp
@@ -177,7 +177,8 @@ private:
             const Norm& norm) const noexcept
         {
             using Scalar = Demote_t<typename Norm::Scalar>;
-            return nanobind::float_{Scalar(norm.clip())};
+            return nanobind::float_{
+                static_cast<double>(Scalar(norm.clip()))};
         }
 
         template<class Norm, std::enable_if_t<!HasClip_v<Norm>>* = nullptr>
@@ -203,7 +204,8 @@ private:
             [[maybe_unused]] const Norm& norm) const noexcept
         {
             using Scalar = Demote_t<typename Norm::Scalar>;
-            return nanobind::float_{Scalar(norm.regularization())};
+            return nanobind::float_{
+                static_cast<double>(Scalar(norm.regularization()))};
         }
 
         template<class Norm,
